+++ src/dataflow/src/render/reduce.rs
@@ -1897,7 +1897,8 @@ fn reduction_type(func: &AggregateFunc) -> ReductionType {
         | AggregateFunc::JsonbObjectAgg { .. }
         | AggregateFunc::ArrayConcat { .. }
         | AggregateFunc::ListConcat { .. }
-        | AggregateFunc::StringAgg { .. } => ReductionType::Basic,
+        | AggregateFunc::StringAgg { .. }
+        | AggregateFunc::RowNumber { .. } => ReductionType::Basic,
     }
 }
 
@@ -2016,7 +2017,8 @@ pub mod monoids {
             | AggregateFunc::JsonbObjectAgg { .. }
             | AggregateFunc::ArrayConcat { .. }
             | AggregateFunc::ListConcat { .. }
-            | AggregateFunc::StringAgg { .. } => None,
+            | AggregateFunc::StringAgg { .. }
+            | AggregateFunc::RowNumber { .. } => None,
         }
     }
 }