+++ src/expr/src/scalar/func.rs
@@ -34,6 +34,7 @@ use lowertest::MzEnumReflect;
 use ore::collections::CollectionExt;
 use ore::fmt::FormatBuffer;
 use ore::result::ResultExt;
+use ore::soft_assert;
 use ore::str::StrExt;
 use pgrepr::Type;
 use repr::adt::array::ArrayDimension;
@@ -5751,7 +5752,7 @@ impl VariadicFunc {
             }
             ArrayToString { .. } => ScalarType::String.nullable(true),
             ListCreate { elem_type } => {
-                debug_assert!(
+                soft_assert!(
                     input_types.iter().all(|t| t.scalar_type.base_eq(elem_type)),
                     "Args to ListCreate should have types that are compatible with the elem_type"
                 );