+++ src/sql/src/plan/lowering.rs
@@ -45,9 +45,9 @@ use ore::collections::CollectionExt;
 use repr::RelationType;
 use repr::*;
 
-use crate::plan::expr::HirRelationExpr;
-use crate::plan::expr::HirScalarExpr;
-use crate::plan::expr::{AggregateExpr, ColumnOrder, ColumnRef, JoinKind};
+use crate::plan::expr::{
+    AggregateExpr, ColumnOrder, ColumnRef, HirRelationExpr, HirScalarExpr, JoinKind, WindowExprType,
+};
 use crate::plan::transform_expr;
 
 /// Maps a leveled column reference to a specific column.
@@ -701,6 +701,149 @@ impl HirScalarExpr {
                 );
                 SS::Column(inner.arity() - 1)
             }
+            Windowing(expr) => {
+                // - For Scalar window functions we need to put a FlatMap operator on top of inner
+
+                let partition = expr.partition;
+                let order_by = expr.order_by;
+
+                match expr.func {
+                    WindowExprType::Scalar(func) => {
+                        *inner = inner
+                            .take_dangerous()
+                            .let_in(id_gen, |id_gen, mut get_inner| {
+                                let order_by = order_by
+                                    .into_iter()
+                                    .map(|o| {
+                                        o.applied_to(id_gen, col_map, &mut get_inner, subquery_map)
+                                    })
+                                    .collect_vec();
+
+                                // Record input arity here so that any group_keys that need to mutate get_inner
+                                // don't add those columns to the aggregate input.
+                                let input_arity = get_inner.typ().arity();
+                                let mut group_key = Vec::new();
+                                for p in partition {
+                                    let key =
+                                        p.applied_to(id_gen, col_map, &mut get_inner, subquery_map);
+                                    if let expr::MirScalarExpr::Column(c) = key {
+                                        group_key.push(c);
+                                    } else {
+                                        get_inner = get_inner.map(vec![key]);
+                                        group_key.push(get_inner.arity() - 1);
+                                    }
+                                }
+
+                                get_inner.let_in(id_gen, |_id_gen, get_inner| {
+                                    let to_reduce = get_inner;
+                                    let input_type = to_reduce.typ();
+                                    let fields = input_type
+                                        .column_types
+                                        .iter()
+                                        .take(input_arity)
+                                        .map(|t| (ColumnName::from("?column?"), t.clone()))
+                                        .collect_vec();
+                                    let agg_input = expr::MirScalarExpr::CallVariadic {
+                                        func: expr::VariadicFunc::RecordCreate {
+                                            field_names: fields
+                                                .iter()
+                                                .map(|(name, _)| name.clone())
+                                                .collect_vec(),
+                                        },
+                                        exprs: (0..input_arity)
+                                            .map(|column| expr::MirScalarExpr::Column(column))
+                                            .collect_vec(),
+                                    };
+                                    let record_type = ScalarType::Record {
+                                        fields,
+                                        custom_oid: None,
+                                        custom_name: None,
+                                    };
+                                    let agg_input = expr::MirScalarExpr::CallVariadic {
+                                        func: expr::VariadicFunc::ListCreate {
+                                            elem_type: record_type.clone(),
+                                        },
+                                        exprs: vec![agg_input],
+                                    };
+                                    let mut agg_input = vec![agg_input];
+                                    agg_input.extend(order_by.clone());
+                                    let agg_input = expr::MirScalarExpr::CallVariadic {
+                                        func: expr::VariadicFunc::RecordCreate {
+                                            field_names: (0..1)
+                                                .map(|_| ColumnName::from("?column?"))
+                                                .collect_vec(),
+                                        },
+                                        exprs: agg_input,
+                                    };
+                                    let list_type = ScalarType::List {
+                                        element_type: Box::new(record_type),
+                                        custom_oid: None,
+                                    };
+                                    let agg_input_type = ScalarType::Record {
+                                        fields: std::iter::once(&list_type)
+                                            .map(|t| {
+                                                (
+                                                    ColumnName::from("?column?"),
+                                                    t.clone().nullable(false),
+                                                )
+                                            })
+                                            .collect_vec(),
+                                        custom_oid: None,
+                                        custom_name: None,
+                                    }
+                                    .nullable(false);
+                                    let func = func.into_expr();
+                                    let aggregate = expr::AggregateExpr {
+                                        func,
+                                        expr: agg_input,
+                                        distinct: false,
+                                    };
+                                    let mut reduce = to_reduce
+                                        .reduce(group_key.clone(), vec![aggregate.clone()], None)
+                                        .flat_map(
+                                            expr::TableFunc::UnnestList {
+                                                el_typ: aggregate
+                                                    .func
+                                                    .output_type(agg_input_type)
+                                                    .scalar_type
+                                                    .unwrap_list_element_type()
+                                                    .clone(),
+                                            },
+                                            vec![expr::MirScalarExpr::Column(group_key.len())],
+                                        );
+                                    let record_col = reduce.arity() - 1;
+
+                                    // Unpack the record
+                                    for c in 0..input_arity {
+                                        reduce = reduce.take_dangerous().map(vec![
+                                            expr::MirScalarExpr::CallUnary {
+                                                func: expr::UnaryFunc::RecordGet(c),
+                                                expr: Box::new(expr::MirScalarExpr::CallUnary {
+                                                    func: expr::UnaryFunc::RecordGet(1),
+                                                    expr: Box::new(expr::MirScalarExpr::Column(
+                                                        record_col,
+                                                    )),
+                                                }),
+                                            },
+                                        ]);
+                                    }
+
+                                    // Append the column with the result of the window function.
+                                    reduce = reduce.take_dangerous().map(vec![
+                                        expr::MirScalarExpr::CallUnary {
+                                            func: expr::UnaryFunc::RecordGet(0),
+                                            expr: Box::new(expr::MirScalarExpr::Column(record_col)),
+                                        },
+                                    ]);
+
+                                    let agg_col = record_col + 1 + input_arity;
+                                    reduce.project((record_col + 1..agg_col + 1).collect_vec())
+                                })
+                            });
+                        SS::Column(inner.arity() - 1)
+                    }
+                }
+            }
         }
     }
 
@@ -826,7 +969,7 @@ impl HirScalarExpr {
                 then: Box::new(then.lower_uncorrelated()?),
                 els: Box::new(els.lower_uncorrelated()?),
             },
-            Select { .. } | Exists { .. } | Parameter(..) | Column(..) => {
+            Select { .. } | Exists { .. } | Parameter(..) | Column(..) | Windowing(..) => {
                 bail!("unexpected ScalarExpr in uncorrelated plan: {:?}", self);
             }
         })